refactor(markdown): restructure markdown utilities
- Moved markdown parsing logic from utils to a dedicated markdown module, improving organization and maintainability.
- Introduced a new MarkdownParser component to handle Markdown content parsing and metadata extraction.
- Added support for frontmatter parsing, enabling structured metadata extraction from Markdown files.
- Implemented comprehensive tests for Markdown parsing, ensuring robustness and security against XSS vulnerabilities.

These changes enhance the Markdown processing capabilities and improve the overall structure of the codebase.
